:construction: refactored: refactored code that compatable with ddeutil-io.

+++ tests/test_loader.py
@@ -4,7 +4,8 @@ from pathlib import Path
 
 import ddeutil.pipe.loader as ld
 import pytest
-from ddeutil.io.models import Params
+from ddeutil.io.param import Params
+from ddeutil.pipe.workflow import Workflow
 
 
 @pytest.fixture(scope='module')
@@ -28,10 +29,11 @@ def params(
             },
         }
     )
-    shutil.rmtree(test_path / ".cache")
+    if (test_path / ".cache").exists():
+        shutil.rmtree(test_path / ".cache")
 
 
-def test_base_loader(params):
+def test_loader(params: Params):
     load: ld.BaseLoad = ld.BaseLoad.from_register(
         name="demo:conn_local_file",
         params=params,
@@ -48,3 +50,37 @@ def test_base_loader(params):
         }
         == load.data
     )
+
+
+def test_simple_loader(params: Params):
+    load = ld.SimLoad(
+        name="conn_local_file",
+        params=params,
+        externals={},
+    )
+    assert {
+        'type': 'connection.LocalFileStorage',
+        'endpoint': 'file:///null/tests/examples/dummy'
+    } == load.data
+
+
+def test_simple_loader_workflow_run_py(params: Params):
+    load = ld.SimLoad(
+        name='run_python_local',
+        params=params,
+        externals={},
+    )
+    assert load.type == Workflow
+    param: str = 'Parameter'
+    g = {'x': param}
+    exec(
+        load.data.get('jobs')[0].get('demo_run').get('stages')[0].get('run'),
+        g,
+    )
+
+    exec(
+        load.data.get('jobs')[0].get('demo_run').get('stages')[1].get('run'),
+        g,
+    )
+
+    assert 1 == g['x']
